Act Bid
Most fields are finished.  Had to stop for the night - need to add a
history selection for the various 'have you performed at?' questions.
Also made updates to graphics - added a ? image, and an icon.  Updated
model a bit to accommodate new form for act bid and to change
"performer" to "bio" if it's going to be reused with teachers.