Now with Class Bid rudiments and refactoring
Refactored so there is just one bid form, and just one view for it so
as not to wound Jon in his soul. :)  Also, it was better design.  Now
there are 2 model forms using this.

to do - work in supporting tables for these two forms, and get the form
intro text dynamic.