Crafted flexible display for bid related info
It’s done for Acts only, but a reasonable format - the view asks the
Act what to display for columns, the Act does the query and presents
the values as tuples along with header info so that business logic is
isolated in the Act in the Model.  The bid Info table will roll with
whatever number of columns and header info it is given.  The only
criteria is that the # of bid info columns and the header labels should
be equal or the table will be messed up.