Added new privilege system to special views. Will add details on priv system to wiki, but in brief: a user has privileges based on their user object's Group membership. We are using none of the actual privilege system, so Groups are simply a marker - we may switch away from abusing the django permission system in this manner, but for now that's where we are.
It is important to note that groups in this implementation live in the database  - this means that any DB refresh will force us to re-enter the groups, and to get them spelled correctly, or
everything will go tits-up. This is brittle, and it's one thing I don't like about this model, but that's where we're at. We really need some way to back up static DB contents.