closed #114
Updating and retesting:

1, 4, 5. 6, 7, 8, 9, 11 -  was done already
2 - Short Description and/or Blurb is gone completely.  I verified that
this
3 - was broken when max_size was changed to max_enrollment
10 - put "label" in manually overridden form widget.

@jonkiparsky  - I beg of you, when you change model names or override a
form element with a different widget, please put on your check list
that you update the hooks to the gbe_form_text help text and label
arrays.  I have found this case more than 3 times in the last hour, and
it would have saved me an hour of debugging time or more, and @burlexpo
some number of hours verifying the fix.  I can see you had actually
done 100% of the work on this fix in the first place, but the
regressive breakages caused the reopening.