Create a way to show draft-required field
Added display to form views that shows that when this is a
draft-ennabled form, fields that are bold are required, even for a
draft.  Then created an array that the Class bid provides for that is a
draft-required field and passed it through the view.

The model is updated, but should not need DB refresh.

May be the end of drafts for classes.  Needs end to end test.