Description is required for Act submit
And there is something completely bizarre about how description works
with respect to the Meta ‘required’ field on Forms.  In this case, it
decided it was not required, even when spec’d as required.  In another
case (ClassBid) the opposite was true.

I have given up trying to understand the design logic, and I’m just
making it work.