Class Bid labels
Fixed labels formatting to stay consistent regardless of the error (or
at least all the error cases I can find).

I don’t love that this is slightly brittle - it reuses model features
that are used by the forms themselves, but the forms have had to
override so many model features that this is not a finite answer for
which elements are actually required - it means “required” settings
have to change in two places.