Basic frame for vendor drafts
Vendors are treated as draftable bids, but we had no navigational
framework for them, it was a super simple post-and-forget type
workflow.  We’ll have bunches of never-fixed vendor drafts if we don’t
give a shared place to work here.

Based on Acts and Classes, with the Act logic in place for submission.