Added error handling for "empty" profile
The profile is not actually empty, the first load of update_profile
creates an empty profile with only a link to the user table.  However,
the user can navigate off without updating which results in a ugly home
page and no purchase_email (or display_name).  Added a check and an
alert for that to the alerts system and tested it.