Close #145
Found the django solution from the django docs - instead of using the
default “id” for the primary key, each parent class has a manually
specified primary key.  Had to go all the way up to scheduler.EventItem
to get it to work, but the problem is fixed.

Also - added a way of describing EventItem so it has some display help
in the /admin pages.  However, for classes the .title isn’t mapping to
anything - this may be because of the multi-inheritance fun… will play
more, but I see it as minor.