Joy!
The EventId in admin now shows some really useful stuff, and I fixed a
bug in class event payload - the “title” is ambiguous because Biddable
and Event both use it, so if the Biddable object wasn’t loaded, this
wouldn’t work.  It matters with the new manager and subclass fetching.

Demoed in - /admin/scheduler/eventitem