Fix for a problem seen in production
the eventitem_id and event_id can be different.  The bug showed up when
that happened, and the self-description in str and unicode then threw
exceptions due to incorrect mapping.  On top of that, it’s possible to
have childless event items (not sure how that happened in prod either)
and that blows up lists of event items.  Added proper error handling
for no-child EventItems and fixed the query.