old bug, just found
If you are in a troupe, you can see the act bid, but you can’t;
- successfully submit an edit, because you are not the contact
- view the act as view-only, because you are not the contact.

I can see act details being useful to a troupe member, but more likely
when it’s scheduled in a show an in connection to rehearsal times, and
such.  I figured that here in the proposal section, a removal from
display would work - that leaves the act info intact when querying the
profile, so that we can use it later for other things.