Framework for user admin
I grabbed JPK’s original work on an Admin look at profiles, which had
fallen out of date, and made a list menu for it, and then updated it to
work like the user update-profile does, except it also has the purchase
email setting available (and required).

That’s the beginning.  We need:
- password reset
- a view on tickets purchased
- a view on schedule
- a massive list of everything - may be same table we have here, may
grow out from that, because it will be a beast of a list — ticket
details, commitments to all parts of conference, and anything else we
can scrape together.