Add Max Volunteer field to form.
Added max volunteer to scheduler.event so it can be varied from booking
to booking (a big win for me in registration) since different
volunteers may be needed at different times of day.

Added it to the view of events table, and to the event scheduling form.

Also fixed a big in the day entry.