No worker = not an error
If the worker is changed to no entry, then the field is skipped in
processing.  It’s a non-error.

If we go with the idea that all volunteer allocations are cleaned prior
to processing (not done yet - @jonkiparsky is working on it) then this
means that any prior worker who is set to “——“ will be deleted.