django-cms up, not working
I got my local install to get it to actually setup using all these
updates.  However, my javascript menu for the editing toolbar will not
pull down - which is the same problem that I had before with javascript
menus on my server… I fear it may be my local, and since I got it up
with no breaks, I’m going to move on to trying on web factional next.