not a better editor, but a better file set
This is a replacement for the django defaults of file manager, links,
images (i.e. picture), teaser and video.  It is brutal and will require
a database resync, as I removed the redundant alternate paths… but it’s
quite a bit nicer to use.

Also - I found that images do not align the way Scratch wanted because
they assumed that they had CSS that wasn’t there… so I fixed this
manually for the imager that I’ve included in this checkin.