Half assed background manager
Saving my current state, but this background manager is half-assed and
does not deserve a pull request until I can find one that is fully
assed, or figure out how to use this one better.

Problems:
- won’t do the <body> style, will only do divs w/in the body.
- only really does background - color or image, doesn’t do other
important parts of <div> tags like margin, border, padding, etc.
- can’t do a “static” version - meaning one would have to edit every
single stinking page to change anything.

Freezing my status so as to not loose work and give a clue what I’m up
to, but this one is not recommended.