1 small thing + pep8
gbe/views.py is now w/out pep8 errors.
That is sort of addictive.

One small change of functionality - the purchase email should not show
up on the regular user’s form.  We use this to connect tickets when a
user has purchased tickets out of the blue on a completely random email
address.  Providing it to the user themselves would  let them snag
someone else’s tickets - so we don’t allow them to edit.

Removed purchase_email from the form and all works well.