Found bug -
When the viewer is a member of a troupe (example - Betty Blaize is a
member of the babydolls), and tries to clone the troupe’s old act from
last year, she fails with a forbidden (perfect!), but the link letting
her clone is shown on the historical bids.  I swiped the clone link
when viewing old acts, since that’s where this comes up.