3/3... sort of
I’m choosing not to mess with scheduler, right now it is free of any
need to check on conference.  i don’t want to unnecessarily link it in
admin and I don’t think anyone but me uses scheduler admin.

Small additions to gbe admin.
Small addition to ticketing.