Fixed it.
Now testing is complete - it tests:
- that a current scheduled class is shown in regular mode
- that a past class is shown in historic mode
- that a current scheduled volunteer opportunity is shown in regular
mode
- that a past scheduled volunteer opportunity is shown in historic mode

And also vice, versa that things that are in regular mode are NOT in
historic mode and things in historic mode are NOT in regular mode.

Of note- getting the timestamps right is super nitpicky and I couldn’t
emulate it perfectly, but this should demonstrate that the timestamp is
not completely wrong.

Also - I chose not to test act assignment - it is one more logical
thread, but it was just one too many, and is not part of the code
changes that I affected in the update.