Better JS
The problem was in the JS, it was way too primitive.  I found a better
example to build off.  This closes the menus when you change focus to a
new menu.  It also persists state - so if you visited a sub menu and
then revisit it, it will persist what you had opened.  Not sure if
that’s a bug or a feature, but it doesn’t strike me as necessarily
“broken”.