Reduce complexity
… and tests.  During runtime, the idea of trying to do a checklist for
a “purchaser” (someone who bought tickets with BPT and who didn’t ever
have a user account), was not useful.  And there was unnecessary
complexity in the get_checklist_items functions, that made the display
ugly and weird.

This reduced the number of feasible unit tests, as some unit tests were
predicated on the different argument patterns that have now been
eliminated.