the rest of contact_by_role
Teacher contact is annoyingly fussy - the class could be bid for
teaching by a teacher that withdrew, but we gave it to an alternate who
is active in the system.  Since we contact both bidders and current
teachers, we should be only avoiding the inactive teacher, regardless
of the conditions.